requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
aniso8601==3.0.2          # via graphene
argon2-cffi==18.3.0
asn1crypto==0.24.0        # via cryptography
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
babel==2.6.0
billiard==3.5.0.4         # via celery
bleach==2.1.4
boto3==1.7.79
botocore==1.10.79         # via boto3, s3transfer
braintree==3.47.0         # via django-payments
cairocffi==0.9.0          # via cairosvg, weasyprint
cairosvg==2.1.3
celery[redis]==4.2.1
certifi==2018.8.13        # via requests
cffi==1.11.5              # via argon2-cffi, cairocffi, cryptography, weasyprint
chardet==3.0.4            # via requests
cryptography==2.3.1       # via django-payments
cssselect2==0.2.1         # via cairosvg, weasyprint
defusedxml==0.5.0         # via cairosvg, python3-openid, social-auth-core
dj-database-url==0.5.0
dj-email-url==0.1.0
django-appconf==1.0.2     # via django-measurement
django-babel==0.6.2
django-bootstrap4==0.0.6
django-cache-url==3.0.0
django-celery-results==1.0.1
django-countries==5.3.1
django-debug-toolbar-request-history==0.0.7
django-debug-toolbar==1.9.1
django-elasticsearch-dsl==0.5.0
django-environ==0.4.5     # via django-graphql-jwt
django-filter==2.0.0
django-fsm==2.6.0
django-graphql-jwt==0.1.10
django-impersonate==1.3
django-js-asset==1.1.0    # via django-mptt
django-measurement==3.0.0
django-mptt==0.9.1
django-payments==0.13.0
django-phonenumber-field==2.0.0
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-prices==1.0.4
django-recaptcha==1.4.0
django-redis==4.9.0
django-render-block==0.5  # via django-templated-email
django-storages==1.6.6
django-templated-email==2.2.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
django==2.0.8
docutils==0.14            # via botocore
elasticsearch-dsl==5.4.0
elasticsearch==5.5.3
faker==0.9.0
freezegun==0.3.10
google-i18n-address==2.3.2
google-measurement-protocol==1.0.0
graphene-django==2.2.0
graphene==2.1.3           # via django-graphql-jwt, graphene-django
graphql-core==2.1         # via graphene, graphene-django, graphql-relay
graphql-relay==0.4.5      # via graphene
html5lib==1.0.1           # via bleach, weasyprint
idna==2.7                 # via cryptography, requests, yarl
jmespath==0.9.3           # via boto3, botocore
jsonfield==2.0.2          # via django-prices-vatlayer
kombu==4.2.1              # via celery
markdown==2.6.11
maxminddb-geolite2==2018.703
maxminddb==1.4.1
measurement==2.0.1        # via django-measurement
more-itertools==4.3.0     # via pytest
mpmath==1.0.0             # via sympy
multidict==4.3.1          # via yarl
oauthlib==2.1.0           # via requests-oauthlib, social-auth-core
pdfrw==0.4                # via weasyprint
phonenumberslite==8.9.10
pillow==5.2.0             # via cairosvg, django-versatileimagefield
pluggy==0.7.1             # via pytest
prices==1.0.0
promise==2.1              # via graphene, graphene-django, graphql-core, graphql-relay
psycopg2-binary==2.7.5
purl==1.4
py==1.5.4                 # via pytest
pycparser==2.18           # via cffi
pyjwt==1.6.4              # via django-graphql-jwt, django-payments, social-auth-core
pyphen==0.9.4             # via weasyprint
pytest-django==3.4.1
pytest-mock==1.10.0
pytest-vcr==0.3.0
pytest==3.7.1
python-dateutil==2.7.3    # via botocore, elasticsearch-dsl, faker, freezegun
python-dotenv==0.9.1
python3-openid==3.1.0     # via social-auth-core
pytz==2018.5              # via babel, celery, django
pyyaml==3.13              # via vcrpy
raven==6.9.0
redis==2.10.6             # via celery, django-redis
requests-oauthlib==1.0.0  # via social-auth-core
requests==2.19.1
rx==1.6.1                 # via graphql-core
s3transfer==0.1.13        # via boto3
singledispatch==3.4.0.3   # via graphene-django
six==1.11.0               # via argon2-cffi, bleach, cryptography, django-templated-email, elasticsearch-dsl, faker, freezegun, graphene, graphene-django, graphql-core, graphql-relay, html5lib, measurement, more-itertools, promise, purl, pytest, python-dateutil, singledispatch, social-auth-app-django, social-auth-core, vcrpy
social-auth-app-django==2.1.0
social-auth-core==1.7.0   # via social-auth-app-django
sqlparse==0.2.4           # via django-debug-toolbar
stripe==2.8.1             # via django-payments
suds-jurko==0.6           # via django-payments
sympy==1.2                # via measurement
text-unidecode==1.2
tinycss2==0.6.1           # via cairosvg, cssselect2, weasyprint
typing==3.6.4             # via google-measurement-protocol, prices, promise
urllib3==1.23             # via elasticsearch, requests
uwsgi==2.0.17.1 ; platform_system != "Windows"
vcrpy==1.13.0
vine==1.1.4               # via amqp
weasyprint==0.42.3
webencodings==0.5.1       # via html5lib, tinycss2
wrapt==1.10.11            # via vcrpy
xmltodict==0.11.0         # via django-payments
yarl==1.2.6               # via vcrpy
